zendesk#3204: Fix "lastseenexpireafter" 32-bit signed int overflow.
Changelog: title
(cherry picked from commit d68441c51641cbcaeb12c117d31a15ee2dcc6c74)

+++ libpromises/cf3globals.c
@@ -46,7 +46,7 @@ AgentType THIS_AGENT_TYPE; /* GLOBAL_C, initialized later */
 /* Internal data structures                                                  */
 /*****************************************************************************/
 
-int LASTSEENEXPIREAFTER = SECONDS_PER_WEEK; /* GLOBAL_P */
+long LASTSEENEXPIREAFTER = SECONDS_PER_WEEK; /* GLOBAL_P */
 
 /*****************************************************************************/
 /* Compatibility infrastructure                                              */
